refactor(vta-service): extract vta-keyspaces + vta-vault (decomposition pilot)
`vta-service` is a 114k-line crate — 44% of the workspace and one compile unit,
so any library change recompiles all of it, with no enforced boundary between
conceptually independent subsystems. This begins decomposing it within the
workspace into subsystem crates. Pilot-first: prove the crate mechanics and the
narrow-dependency seam on the cleanest leaf, then repeat per subsystem.

Phase 0 — new `vta-keyspaces`: the 29 keyspace-name constants + the backup
partition, a dependency-free leaf so subsystem crates can name keyspaces without
depending on `vta-service`. `vta-service::keyspaces` re-exports it (every
`crate::keyspaces::*` reference unchanged); the `no_bare_keyspace_literals`
guard stays in `vta-service` since it must scan that crate's source.

Phase 1 — new `vta-vault`: the ~8.7k-line holder credential vault (storage,
model, query, receive/verify, present, status, bbs, di_verify, consent),
extracted from `vta-service/src/vault/`. It takes narrow dependencies
(KeyspaceHandle, ActScope, resolver args) and never AppState — which is why it
was the clean pilot. Re-exported as `vta-service::vault`, so the dispatch
handlers, sweeper, and credential_exchange keep their `crate::vault::…` paths.
`bbs` / `webvh` features wire through to `vta-vault/bbs` / `vta-vault/webvh`.

Pure extraction, no behaviour change: files moved as git renames (history
preserved); 103 vault tests run in the new crate; the dispatch and
credential-custody (#776) tests still run in vta-service and exercise the seam.
vta-enclave (the one external consumer of vta-service internals) is unaffected.
vta-service drops from ~113.9k to ~104.9k source lines.

Frictions were exactly what extraction surfaces: a few feature/dep edges
(async-trait, multibase, vta-sdk/client under webvh) and one pub(crate)->pub
widening (resolve_di_issuer_key) where a fn now crosses a real crate boundary.

cargo test -p vta-keyspaces -p vta-vault --features bbs,webvh -p vta-service:
1255 passed, 0 failed. Workspace + enclave build; clippy + fmt clean.
