+++ Cargo.toml
@@ -6,9 +6,11 @@ members = [
   "pnm-cli",
   "tests/e2e",
   "vta-cli-common",
+  "vta-keyspaces",
   "vta-mcp",
   "vta-mobile-core",
   "vta-sdk",
+  "vta-vault",
   "vta-service",
   "vta-enclave",
   "vtc-client",
@@ -27,9 +29,11 @@ default-members = [
   "pnm-cli",
   "tests/e2e",
   "vta-cli-common",
+  "vta-keyspaces",
   "vta-mcp",
   "vta-sdk",
   "vta-service",
+  "vta-vault",
   "vta-enclave",
   "vtc-client",
   "vtc-service",
