+++ vta-service/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "vta-service"
 description = "Service for Verifiable Trust Agents operating in Verifiable Trust Communities"
-version = "0.12.25"
+version = "0.12.26"
 edition.workspace = true
 publish.workspace = true
 authors.workspace = true
@@ -22,13 +22,13 @@ tsp = [
     "dep:affinidi-messaging-sdk",
     "affinidi-messaging-sdk/tsp",
 ]
-webvh = ["dep:didwebvh-rs", "dep:url", "dep:reqwest"]
+webvh = ["dep:didwebvh-rs", "dep:url", "dep:reqwest", "vta-vault/webvh"]
 setup = ["webvh", "dep:tempfile"]
 # BBS+ (`bbs-2023`) credential support in the vault — pulls in the BLS12-381
 # curve (`affinidi-bbs`) + the `bbs-2023` Data-Integrity cryptosuite. Off by
 # default; the holder/verifier paths (receive, present, verify) build behind
 # it. Issuer signing stays audit-gated (not exposed from the VTA).
-bbs = ["dep:affinidi-bbs", "affinidi-data-integrity/bbs-2023"]
+bbs = ["dep:affinidi-bbs", "affinidi-data-integrity/bbs-2023", "vta-vault/bbs"]
 # Enables the offline CLI command helpers (bootstrap_cli, webvh_cli) that
 # synthesize super-admin `AuthClaims` without wire-level verification. The
 # local `vta` binary enables this; the enclave binary does not (its
@@ -93,6 +93,11 @@ vti-common = { path = "../vti-common", version = "0.11", features = ["encryption
 # from this crate into a shared crate so external integrations can reuse them).
 # Per-backend features are activated by this crate's matching features below.
 vti-secrets = { path = "../vti-secrets", version = "0.1" }
+vta-keyspaces = { path = "../vta-keyspaces", version = "0.1" }
+# Holder credential vault (storage/query/receive/present/status), extracted to
+# its own crate and re-exported as `crate::vault`. Feature edges: this crate's
+# `bbs` / `webvh` features enable `vta-vault/bbs` / `vta-vault/webvh`.
+vta-vault = { path = "../vta-vault", version = "0.1" }
 vta-sdk = { path = "../vta-sdk", version = "0.19", features = ["didcomm", "sealed-transfer", "provision-integration", "openapi", "crypto-provider", "acl-setup"] }
 # DID-VM-resolved WebAuthn assertion verifier — drives the new
 # `vta/auth/passkey-login-{start,finish}/1.0` trust-task handlers.
