+++ vta-vault/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "vta-vault"
+description = "VTA holder credential vault — storage, query, receive/verify, present, and status refresh for credentials a holder stores on a VTA"
+readme = "README.md"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+authors.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[features]
+default = []
+# BBS+ (`bbs-2023`) selective-disclosure credentials in the vault — the
+# receive/present paths for pseudonym credentials. Pulls in the BLS12-381 curve
+# and the `bbs-2023` Data-Integrity cryptosuite. Off by default.
+bbs = ["dep:affinidi-bbs", "affinidi-data-integrity/bbs-2023"]
+# Holder-side status refresh (`HttpStatusListResolver`) — fetches an issuer's
+# BitstringStatusList over HTTP to re-check a held credential's validity.
+# Gated because it needs the (SSRF-hardened) foreign-fetch HTTP client, which
+# lives behind vta-sdk's `client` feature (`vta_sdk::http`).
+webvh = ["dep:reqwest", "vta-sdk/client"]
+
+[dependencies]
+vta-keyspaces = { path = "../vta-keyspaces", version = "0.1" }
+vti-common = { path = "../vti-common", version = "0.11", features = ["encryption"] }
+vta-sdk = { path = "../vta-sdk", version = "0.19", features = ["didcomm", "sealed-transfer"] }
+affinidi-crypto = { workspace = true }
+affinidi-secrets-resolver = { workspace = true }
+affinidi-data-integrity = { workspace = true }
+affinidi-sd-jwt = { workspace = true }
+affinidi-sd-jwt-vc = { workspace = true }
+affinidi-status-list = { workspace = true }
+affinidi-did-resolver-cache-sdk = { workspace = true }
+affinidi-bbs = { workspace = true, optional = true }
+async-trait = "0.1"
+serde = { workspace = true }
+serde_json = { workspace = true }
+base64 = { workspace = true }
+multibase = { workspace = true }
+ed25519-dalek = { workspace = true }
+chrono = { workspace = true }
+uuid = { workspace = true }
+tracing = { workspace = true }
+reqwest = { workspace = true, optional = true }
+
+[dev-dependencies]
+tokio = { workspace = true }
+tempfile = "3"
