+++ vta-vault/src/index.rs
