+++ vta-vault/src/lib.rs
